requirements.txt
asyncpraw==7.3.0
discord.py==1.7.2
pylunar>=0.6.0